Minor fixes & improvements

+++ nd4j-backends/nd4j-backend-impls/nd4j-native/src/main/java/org/nd4j/linalg/cpu/nativecpu/blas/CpuLevel1.java
@@ -69,7 +69,7 @@ public class CpuLevel1 extends BaseLevel1 {
             // non-EWS dot variant
             Dot dot = new Dot(X, Y);
             Nd4j.getExecutioner().exec(dot);
-            return dot.getFinalResult().floatValue();
+            return dot.getFinalResult().doubleValue();
         }
     }
